+++ be/src/main/java/com/codesquad/airbnb/naverapi/service/GeoCodeService.java
@@ -1,9 +1,10 @@
 package com.codesquad.airbnb.naverapi.service;
 
-import com.codesquad.airbnb.naverapi.web.dto.Address;
-import com.codesquad.airbnb.naverapi.web.dto.GeoCodeResponseDto;
+import com.codesquad.airbnb.common.util.NaverAPIConnector;
 import com.codesquad.airbnb.naverapi.ApiKey;
 import com.codesquad.airbnb.naverapi.domain.GeoCode;
+import com.codesquad.airbnb.naverapi.web.dto.Address;
+import com.codesquad.airbnb.naverapi.web.dto.GeoCodeResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.json.simple.parser.ParseException;
@@ -17,9 +18,9 @@ import org.springframework.transaction.annotation.Transactional;
 public class GeoCodeService {
     private final ApiKey apiKey;
 
-    public GeoCodeResponseDto convertAddressToCoordinate(Address address) throws ParseException {
+    public GeoCodeResponseDto convertAddressToGeoCode(Address address) throws ParseException {
         String addressString = address.toValue();
-        GeoCode geoCode = GeoCode.getGeoCodeFromNaverApi(addressString, apiKey);
+        GeoCode geoCode = NaverAPIConnector.convertToGeoCode(addressString, apiKey);
         log.info("[Accomm. Service] (latitude, longitude): ({}, {})",
                 geoCode.getLatitude(), geoCode.getLongitude());
         return new GeoCodeResponseDto(geoCode.getLatitude(), geoCode.getLongitude());
